Additional invalid size diagnostics
Previously, when an invalidly sized download was encountered, only the
sizes were logged.  This change adds additional debug diagnostic
information that prints the actual contents of the downloaded files.
This is only a temporary change for issue #53, as it would wreak havoc
if enabled any time someone enabled debug logging.